requirements.txt
fastapi
uvicorn
openai
pandas
requests
python-dotenv
